Fixed leaked tracer
Problem was ODEs expect a callable (t,) -> u,
but an AbstractControl can't be passed as that type without violating purity, as long as we are jitting ODEs directly.
Hence, use eqx.partition and combine to pass the jit barrier.
It might also work to avoid the jit in the first place, will try that next.